+++ src/forms/postForm.tsx
@@ -1,0 +1,12 @@
+import { FormState } from './Form'
+
+
+export function postForm(route: string, formData: FormState) {
+  return fetch(route, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(formData),
+  })
+}
